services/stt-service/requirements.txt
numpy~=2.4.3
soundfile~=0.13.1
fastapi~=0.135.1
faster-whisper~=1.2.1
python-multipart~=0.0.22
uvicorn~=0.41.0
python-dotenv~=1.2.2
